+++ tools/publish-aos.mjs
@@ -8,8 +8,6 @@ import version from '../version.mjs';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-const srcCodeTxIdPlaceholder = '__INSERT_SOURCE_CODE_ID__';
-
 const bundledLua = fs.readFileSync(
   path.join(__dirname, '../dist/aos-bundled.lua'),
   'utf-8',
@@ -17,12 +15,15 @@ const bundledLua = fs.readFileSync(
 
 const dryRun =
   process.argv.includes('--dry-run') || process.env.DRY_RUN === 'true';
+
 const walletPath = process.argv.includes('--wallet-file')
   ? process.argv[process.argv.indexOf('--wallet-file') + 1]
   : process.env.WALLET_PATH || path.join(__dirname, 'key.json');
+
 const jwk = process.env.WALLET
   ? JSON.parse(process.env.WALLET)
   : JSON.parse(fs.readFileSync(walletPath, 'utf-8'));
+
 const signer = new ArweaveSigner(jwk);
 const turbo = TurboFactory.authenticated({ signer });
 const publishingTags = Object.entries({
@@ -36,38 +37,18 @@ const publishingTags = Object.entries({
   .filter(([_, value]) => value !== undefined)
   .map(([name, value]) => ({ name, value }));
 
-/**
- * NOTE: with the current use of SOURCE-CODE-TX-ID, we have to publish the generate the source code twice
- * to get the tx id into the bundled file. In the future, we should move to using incremental
- * bundle versions to avoid this issue.
- */
-const data1 = createData(bundledLua, signer, {
+const data = createData(bundledLua, signer, {
   tags: publishingTags,
 });
-await data1.sign(signer);
+await data.sign(signer);
 
-console.log('Generated source code data item with id: ' + data1.id);
-// replace placeholder with actual tx id
-const bundledLuaWithTxId = bundledLua.replace(srcCodeTxIdPlaceholder, data1.id);
-
-const data2 = createData(bundledLuaWithTxId, signer, {
-  tags: [...publishingTags, { name: 'Original-Tx-Id', value: data1.id }],
-});
-await data2.sign(signer);
-
-console.log('Generated bundled data item with id: ' + data2.id);
+console.log('Generated source code data item with id: ' + data.id);
 
 if (!dryRun) {
   console.log('Publishing ANT Source code to Arweave...');
-  await Promise.all([
-    turbo.uploadSignedDataItem({
-      dataItemSizeFactory: () => data1.getRaw().byteLength,
-      dataItemStreamFactory: () => data1.getRaw(),
-    }),
-    turbo.uploadSignedDataItem({
-      dataItemSizeFactory: () => data2.getRaw().byteLength,
-      dataItemStreamFactory: () => data2.getRaw(),
-    }),
-  ]);
+  await turbo.uploadSignedDataItem({
+    dataItemSizeFactory: () => data.getRaw().byteLength,
+    dataItemStreamFactory: () => data.getRaw(),
+  });
 }
-console.log('Tagged source code tx id: ' + data1.id);
+console.log('Tagged source code tx id: ' + data.id);